rename shd-bleep-object to shd-bleep
