+++ src/main/CMakeLists.txt
@@ -89,7 +89,7 @@ set(shadow_srcs
     host/shd-host.c
     host/shd-network-interface.c
     host/shd-tracker.c
-    host/shd-bleep-object.c
+    host/shd-bleep.c
 
     routing/shd-payload.c
     routing/shd-packet.c
